refactor: split flex layout domain
